v0/requirements.txt
flask>=3.0
pillow>=10.0
pytesseract>=0.3.10
